requirements.txt
torch==2.3.0
numpy==1.26.4
torchvision==0.18.0
pillow==10.3.0
transformers==4.49.0
diffusers==0.30.2
accelerate==0.26.0
tqdm==4.66.4
tensorboard==2.19.0